requirements.txt
Django==5.2.3
djangorestframework==3.16.0
dotenv==0.9.9
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pillow==11.3.0
django-filter==25.1
djangorestframework_simplejwt==5.5.0
